text_to_sql/requirements.txt
streamlit
mysql-connector-python
langchain
openai
python-dotenv
langchain-experimental